feat(oas-to-snippet): plugin architecture
