+++ packages/oas-to-snippet/package.json
@@ -11,17 +11,9 @@
       "require": "./dist/index.cjs",
       "import": "./dist/index.js"
     },
-    "./supportedLanguages": {
-      "require": "./dist/supportedLanguages.cjs",
-      "import": "./dist/supportedLanguages.js"
-    },
-    "./types": {
-      "require": "./dist/types.d.cts",
-      "import": "./dist/types.d.ts"
-    },
-    "./utils": {
-      "require": "./dist/lib/utils.cjs",
-      "import": "./dist/lib/utils.js"
+    "./languages": {
+      "require": "./dist/languages.cjs",
+      "import": "./dist/languages.js"
     },
     "./package.json": "./package.json"
   },
@@ -45,6 +37,7 @@
   "scripts": {
     "attw": "attw --pack --format table-flipped",
     "build": "tsup",
+    "build:docs": "npm run build && node bin/generate-target-markdown-table.js",
     "lint": "npm run lint:types && npm run lint:js",
     "lint:js": "eslint . --ext .js,.ts --ignore-path ../../.gitignore",
     "lint:types": "tsc --noEmit",
@@ -53,9 +46,8 @@
     "test": "vitest run --coverage"
   },
   "dependencies": {
-    "@readme/httpsnippet": "^8.0.1",
-    "@readme/oas-to-har": "file:../oas-to-har",
-    "httpsnippet-client-api": "^7.0.0-beta.0"
+    "@readme/httpsnippet": "^8.1.3",
+    "@readme/oas-to-har": "file:../oas-to-har"
   },
   "devDependencies": {
     "@readme/oas-examples": "^5.12.0",
@@ -63,6 +55,7 @@
     "@types/node": "^20.8.7",
     "@vitest/coverage-v8": "^0.34.6",
     "har-examples": "^3.1.1",
+    "httpsnippet-client-api": "^7.0.0-beta.4",
     "oas": "file:../oas",
     "tsup": "^7.2.0",
     "type-fest": "^4.5.0",
